+++ docs/changelog.rst
@@ -13,6 +13,7 @@ v0.2.0
 - Add support for :ref:`template paths <guide-templating--template-paths>`
 - Add support for :ref:`dynamic generation <guide-dynamic-generation>`
 - Add support for installation of boilerplates from git repositories
+- Add :attr:`Context.state` attribute for propagating stateful information
 
 **Enhancements and Changes**